modify config
