+++ lagent/rag/nlp/__init__.py
@@ -1,11 +1,10 @@
 from .sentence_transformer_embedder import SentenceTransformerEmbedder
 from .tokenizer import SimpleTokenizer
-from .vectore_store import FaissDatabase, DocumentDB
+from .vectore_store import FaissDatabase
 
 
 __all__ = [
     'SentenceTransformerEmbedder',
     'FaissDatabase',
-    'DocumentDB',
     'SimpleTokenizer',
 ]
