+++ lagent/rag/nlp/vectore_store.py
@@ -8,22 +8,7 @@ from typing import List, Tuple, Optional, Union, Callable, Dict, Any
 
 from lagent.rag.nlp.sentence_transformer_embedder import SentenceTransformerEmbedder
 from lagent.utils import create_object
-
-
-class DocumentDB:
-    """
-        Represents a document with an ID, content, and optional metadata.It is used for building database.
-
-        Args:
-            id (str): The unique identifier for the document.
-            content (str): The textual content of the document.
-            metadata (Optional[Dict]): Additional metadata for the document.
-                Defaults to an empty dictionary if not provided.
-        """
-    def __init__(self, id: str, content: str, metadata: Optional[Dict] = None):
-        self.id = id
-        self.content = content
-        self.metadata = metadata or {}
+from ..schema import DocumentDB
 
 
 class FaissDatabase:
