+++ lagent/rag/processors/build_db.py
@@ -1,8 +1,8 @@
 from lagent.rag.schema import MultiLayerGraph
-from lagent.rag.nlp import FaissDatabase, DocumentDB
+from lagent.rag.nlp import FaissDatabase
 from lagent.rag.pipeline import BaseProcessor, register_processor
 from lagent.rag.nlp import SentenceTransformerEmbedder
-from lagent.rag.schema import Chunk, Node
+from lagent.rag.schema import Chunk, Node, DocumentDB
 from lagent.utils import create_object
 
 from typing import Optional, List
@@ -48,10 +48,6 @@ class BuildDatabase(BaseProcessor):
 
         faiss_db = FaissDatabase.from_documents(documents, embedding_function)
 
-        # # save faiss index
-        # faiss_db.save_local(f'{self.db_index_path}_chunks.pkl')
-        # print(f"FAISS index created and saved to {self.db_index_path}_chunks.pkl")
-
         return faiss_db
 
     def initialize_entity_faiss(self, entities: List[Node]) -> FaissDatabase:
@@ -70,9 +66,5 @@ class BuildDatabase(BaseProcessor):
         embedding_function = self.embedder
 
         faiss_db = FaissDatabase.from_documents(documents=documents, embedder=embedding_function)
-
-        # # save faiss index
-        # faiss_db.save_local(f'{self.db_index_path}_entities')
-        # print(f"FAISS index created and saved to {self.db_index_path}_entities.")
 
         return faiss_db
