+++ lagent/rag/utils.py
@@ -26,11 +26,6 @@ def normalize_edge(edge: Relationship) -> Relationship:
         return edge
 
 
-def name_to_id(name: str) -> str:
-    _id = hashlib.md5(name.encode('utf-8')).hexdigest()
-    return _id
-
-
 def tuple_to_str(data: tuple[str, str]) -> str:
     return f"{data[0]},{data[1]}"
 
@@ -44,78 +39,6 @@ def str_to_tuple(data: str) -> tuple[str, str]:
         source = items[0]
         target = items[1]
         return (source, target)
-
-
-def get_id_map_nodes(nodes: List[Node], storage: Optional[Storage], stage: str) -> Dict[str, Node]:
-    load_dict_items = storage.get(f"id_map_nodes_{stage}")
-    id_map_nodes = {}
-    if load_dict_items is None:
-
-        map_items_dict = []
-        for node in nodes:
-            id_map_nodes[node.id] = node
-            map_items_dict.append({node.id: node.to_dict()})
-        storage.put(f"id_map_nodes_{stage}", map_items_dict)
-
-    else:
-        for map_item in load_dict_items:
-            k, v = map_item.popitem()
-            id_map_nodes[k] = Node.dict_to_node(v)
-
-    return id_map_nodes
-
-
-def get_id_map_relationships(relationships: List[Relationship], storage: Optional[Storage], stage: str) -> Dict[
-    str, Relationship]:
-    load_dict_items = storage.get(f"id_map_relationships_{stage}")
-    id_map_relationships = {}
-
-    if load_dict_items is None:
-        map_items_dict = []
-        for edge in relationships:
-            name = tuple_to_str((edge.source, edge.target))
-            id_map_relationships[name] = edge
-            map_items_dict.append({name: edge.to_dict()})
-        storage.put(f"id_map_relationships_{stage}", map_items_dict)
-
-    else:
-        try:
-            for map_item in load_dict_items:
-                k, v = map_item.popitem()
-                id_map_relationships[k] = Relationship.dict_to_edge(v)
-        except ValueError as e:
-            print(f"Caught an exception: {e}")
-
-    return id_map_relationships
-
-
-def create_igraph(nodes: List[Node], edges: Optional[List[Relationship]] = None):
-    # given nodes and edges, create a graph for leidenalg
-    g = ig.Graph(directed=False)
-
-    for node in nodes:
-
-        vertex_attr = {'id': node.id, 'name': node.id}
-        if node.description is not None:
-            vertex_attr['description'] = node.description
-        if hasattr(node, 'degree'):
-            vertex_attr['degree'] = node.degree
-        if hasattr(node, 'community'):
-            vertex_attr['community'] = node.community
-
-        g.add_vertex(**vertex_attr)
-
-    if edges is None:
-        return g
-
-    for edge in edges:
-        edge_attr = edge.to_dict()
-        edge_attr['source'] = edge_attr.get('source')
-        edge_attr['target'] = edge_attr.get('target')
-
-        g.add_edge(**edge_attr)
-
-    return g
 
 
 def filter_nodes_by_commu(nodes: List[Node], community_id: str) -> List[Node]:
